A restructure was done to the initial test-suite.  This is done in preparation for Testing Thursdays!
Matthew Budram
